refactor(views) deduplicate dataplane views
Removed dupicated code from views and moved it to separate,
partial component called `Dataplanes` and all views now are using
this partial component.

Because of that I had to rename `Dataplanes` view to
`AllDataplanes` now
